Final Adjustments!
Had to restructure every file for the HTML, JavaScript, and CSS code
blocks. In regards to the HTML there wasn't much to change however, I
needed to add the "DOMContentLoaded" event listener as well as the
remove function and id's/classes to call back in the JavaScript file.
As for the CSS file, I added media queries for other devices and I
to make an adjustment within my top and bottom classes to prevent the
top and bottom texts from bunching together at the bottom of the image.